Use capacity to determine placement of address bit A8
Previously, this would look at the device type selected and for CAT25040
only, put A8 in the command byte rather than in a second address byte.
However, it seems that all compatible EEPROMs with 4kbit capacity also
use this same exception (e.g. also the ST M95040 or the Microchip
AT25040B).

With this commit, this exception is made for all devices with a capacity
of 4kbit, which makes the code a bit more generic and making it easier
to support more chips in the future (this prepares for fixing #4).