Make specifying devices more flexible
Previously there was a predefined enum of supported devices, that was
translated to a capacity and pageSize value using a switch, which makes
it hard to support devices with different settings.

Now, the device is specified as a struct containing capacity and
pageSize, which makes it trivial to define a new device in the sketch if
it is not yet explicitly supported by the library. By making this struct
have the same name as the original enum type and predefining variables
with the same name as the original enum constants, existing code will
continue to work without changes.

This fixes #4.