Add const to method arguments where appropriate
For the write methods, this makes the buffer argument pointer-to-const,
so callers know that the buffer passed will not be modified (and the
compiler can check this).

Where possible, all argument themselves are also made const (including
the buffer arguments, which are now const-pointer-to-const). This was
already done for most of the arguments, so things are bit more
consistent now.

This fixes #2.