Implement timeout waiting for the ready bit
This prevents an infinite loop when the device is not responding and
only 1 bits are read (e.g. connections problems, not powered, etc.).

The timeout is fixed to 11ms, which is the 10ms max write time with some
margin. Although most chips have max 5ms write time, some (especially
larger ones like M95M02) have 10ms. To allow using all of them, use a
bigger timeout. This waits a bit longer than needed on other chips, but
since this should only be an exception, that is ok.

All methods that have the timeout now return 0 to indicate the timeout.
This makes it impossible to distinguish between timeouts or invalid
parameters, without proper error codes, this is otherwise tricky to
distinguish.

This fixes #5.