Predefine settings for additional devices
This adds the ST M95xxx range, as well as the 2Mbit CAT25M02 that was
still missing.

The M95xxx devices are essentially essential to the CAT25xxx, except for
the M95640, which has a different page size.

These settings were based on the datasheets, only the M95640 was
actually tested.